backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
pdfplumber==0.10.3
pypdf==3.17.1
pandas==2.1.3
numpy==1.26.2
redis==5.0.1
httpx==0.25.1
python-dotenv==1.0.0
alembic==1.12.1
openpyxl==3.1.2
